Fix warning when compiling with -Wconversion
This is not a real issue, but adding the cast to CGFloat silences this warning, which could appear for users of PureLayout if they have some uncommon build settings.

Fixes #73, fixes #49

+++ PureLayout/PureLayout/PureLayout+Internal.h
@@ -29,7 +29,7 @@
 
 /** A constant that represents the smallest valid positive value for the multiplier of a constraint,
     since a value of 0 will cause the second item to be lost in the internal auto layout engine. */
-static const CGFloat kMULTIPLIER_MIN_VALUE = 0.00001; // very small floating point numbers (e.g. CGFLOAT_MIN) can cause problems
+static const CGFloat kMULTIPLIER_MIN_VALUE = (CGFloat)0.00001; // very small floating point numbers (e.g. CGFLOAT_MIN) can cause problems
 
 
 /**
